+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/CheckoutPresenter.java
@@ -515,11 +515,6 @@ public class CheckoutPresenter extends MvpPresenter<CheckoutView, CheckoutProvid
                 }
 
                 @Override
-                public void driveToShowOneTap() {
-                    //donothing
-                }
-
-                @Override
                 public void driveToShowPaymentMethodSelection() {
                     getView().showPaymentMethodSelection();
                 }